When reloading, force kill old process after a timeout
This is related to #1006. If the process does not shutdown with
SIGTERM, use SIGKILL.

Note that the `timeout` argument for `wait()` is an addition for
Python 3. If we still need to support Python 2, the easiest way would
be to just not to do that and instead use `kill()` instead of
`terminate()` to not have this issue.

Also revert the change in `healthcheck.py` checking for the PPID.
AFAIK, the process is not reparented.