Be explicit when accessing an subclass of Family
Many objects subclass Family and the code will often redefine
functions such as extensive() and therefore could lead to
the presentation of the object itself and not the Family class.

The function family() was therefore changed to return a Family
object using the underlying AFI and SAFI.

This will lead to slower performance as the object has to be
created instead of using the underlying afi/safi which were
previously returned.

As the API changed, we know need a new afi_safi() function to
return the tuple previously provided.