answer to #1191
Not all code paths leading to the TCP connection closing
were leading to message being sent to the api.

The peer _reset function was refactored to split to create
a peer._close() function which wrap the api call and FSM state
machine change when calling proto.close().

All previous calls to proto._close() were then wrapped to this
new call, which should resolve the issue reported in #1191.