poetry fun
We do not really use poetry as we vendor our dependencies.
But to make life easier to other devs, we have it so that they can use
virtual env.
Updated the dependency version to stop github from getting scared with
the project.
Trust me .. It's fine .. or trust no one .. I do not care
The diff looks sound.