change extended-encoding for #1209
The code was changed in two patches, one to ensure the ordering
to be consistend when appending to extended communities and not
break the functional testing.

The other change was to merge the feature of add_and_merge into
attribute.add as some code somewhere is using add instead of
add_and_merge and add should do the right thing.

This should be all that is needed to put the issue to rest.